combine Save and Load  bluesprints screens  - remove duplication
